+++ src/components/mypage/reservations/Card/index.ts
@@ -1,0 +1,2 @@
+export { default as ModelReservationCard } from './ModelReservationCard';
+export { default as DesignerReservationCard } from './DesignerReservationCard';
