+++ src/components/mypage/reservations/List/index.ts
@@ -1,0 +1,2 @@
+export { default as ReservationList } from './ReservationList';
+export { default as ReservationListSkeleton } from './ReservationListSkeleton';
