+++ src/components/mypage/reservations/Tabs/index.ts
@@ -1,0 +1,3 @@
+export { default as ReservationTabs } from './ReservationTabs';
+export { default as CategoryChips } from './CategoryChips';
+export { default as MonthDropdown } from './MonthDropdown';
