+++ src/components/mypage/reservations/index.ts
@@ -1,7 +1,3 @@
-export { default as ReservationTabs } from './ReservationTabs';
-export { default as CategoryChips } from './CategoryChips';
-export { default as MonthDropdown } from './MonthDropdown';
-export { default as ModelReservationCard } from './ModelReservationCard';
-export { default as DesignerReservationCard } from './DesignerReservationCard';
-export { default as ReservationList } from './ReservationList';
-export { default as ReservationListSkeleton } from './ReservationListSkeleton';
+export * from './Tabs';
+export * from './Card';
+export * from './List';
